requirements.txt
awscli==1.11.45
beautifulsoup4==4.5.3
boto3==1.4.4
botocore==1.5.8
bottlenose==1.1.2
colorama==0.3.7
dj-database-url==0.4.2
Django==1.10.5
django-registration-redux==1.4
django-storages==1.5.2
docutils==0.13.1
gunicorn==19.6.0
jmespath==0.9.1
lxml==3.7.3
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
pyasn1==0.2.2
python-dateutil==2.6.0
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
whitenoise==3.2.3
